% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_core.R
\name{export_synergy_excel}
\alias{export_synergy_excel}
\title{Export synergy results to a multi-sheet Excel workbook}
\usage{
export_synergy_excel(synergy_res, output_path)
}
\arguments{
\item{synergy_res}{A synergy_result object}

\item{output_path}{Path for the output .xlsx file}
}
\description{
Export synergy results to a multi-sheet Excel workbook
}
