% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_kegg.R
\name{kegg_cache_path}
\alias{kegg_cache_path}
\title{Path to the local KEGG cache file}
\usage{
kegg_cache_path(organism = "hsa")
}
\arguments{
\item{organism}{KEGG organism code.}
}
\value{
Absolute path to the RDS file used by \code{get_kegg_data_cached}.
  The file may not yet exist.
}
\description{
Path to the local KEGG cache file
}
