% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_plot.R
\name{plot_synergy_contrib}
\alias{plot_synergy_contrib}
\title{Bar chart showing effect contribution for top synergistic genes}
\usage{
plot_synergy_contrib(synergy_res, n = 20, direction = "up")
}
\arguments{
\item{synergy_res}{A synergy_result object}

\item{n}{Top N genes to show (default 20)}

\item{direction}{"up" or "down"}
}
\value{
A ggplot object
}
\description{
For UP genes: shows Increase (FC-1) for CvsNT, AvsNT, BvsNT
For DOWN genes: shows Decrease (1-FC)
}
