% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_plot.R
\name{plot_synergy_overlap}
\alias{plot_synergy_overlap}
\title{Venn/Euler diagram of gene overlaps across key comparisons}
\usage{
plot_synergy_overlap(synergy_res)
}
\arguments{
\item{synergy_res}{A synergy_result object}
}
\value{
A ggplot or euler plot object
}
\description{
Venn/Euler diagram of gene overlaps across key comparisons
}
