% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_plot.R
\name{plot_synergy_volcano}
\alias{plot_synergy_volcano}
\title{Volcano plot highlighting synergistic genes}
\usage{
plot_synergy_volcano(synergy_res, highlight_label = 10)
}
\arguments{
\item{synergy_res}{A synergy_result object}

\item{highlight_label}{Number of top synergy genes to label (default 10)}
}
\value{
A ggplot object
}
\description{
Volcano plot highlighting synergistic genes
}
