% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_io.R
\name{read_diff_result}
\alias{read_diff_result}
\title{Read an edgeR or DESeq2 result file}
\usage{
read_diff_result(filepath)
}
\arguments{
\item{filepath}{Path to the result file}
}
\value{
A data.frame with standardized columns
}
\description{
Accepts tab-delimited .xls files produced by edgeR or DESeq2 pipelines.
Required columns: gene_id, gene_name, log2FC, Pvalue, Qvalue, updown
}
