% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynergyAnalysis-package.R
\docType{package}
\name{SynergyAnalysis-package}
\alias{SynergyAnalysis}
\alias{SynergyAnalysis-package}
\title{SynergyAnalysis: identify synergistic genes from transcriptomics data}
\description{
Framework and Shiny app for finding genes where a combination treatment A+B
produces an effect exceeding the sum of A alone and B alone.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/puweilin/SynergyAnalysis}
  \item Report bugs at \url{https://github.com/puweilin/SynergyAnalysis/issues}
}

}
\author{
\strong{Maintainer}: puweilin \email{puweilin@users.noreply.github.com}

}
\keyword{internal}
